renamed index to class name, approve allowance into class
