+++ packages/hedera-wallet-snap/packages/snap/src/rpc/account/approveAllowance.ts
@@ -22,7 +22,7 @@ import { providerErrors } from '@metamask/rpc-errors';
 import { divider, heading, text } from '@metamask/snaps-ui';
 import _ from 'lodash';
 import { HederaClientFactory } from '../../snap/HederaClientFactory';
-import { HederaServiceImpl } from '../../services/impl/hedera';
+import { HederaServiceImpl } from '../../services/impl/hedera/HederaServiceImpl';
 import { SnapUtils } from '../../utils/SnapUtils';
 import { MirrorTokenInfo, TxReceipt } from '../../types/hedera';
 import {
